requirements.txt
ipykernel
pandas
numpy
pyarrow 
sqlalchemy 
psycopg2 
pyreadstat
google-cloud-storage